refactor(user): neutralize user data service filename
