+++ turbo/apps/api/src/signals/routes/cli-auth.ts
@@ -19,7 +19,7 @@ import { issueCliToken$ } from "../services/cli-auth.service";
 import {
   updateUserPreferences$,
   userPreferences,
-} from "../services/zero-user-data.service";
+} from "../services/user-data.service";
 import { isValidTimeZone } from "../utils";
 
 const DEVICE_CODE_CHARS = "ABCDEFGHJKMNPQRSTUVWXYZ23456789";
