+++ turbo/apps/api/src/signals/services/zero-agentphone.service.ts
@@ -63,7 +63,7 @@ import { createUserMessageDocument } from "./zero-chat-user-message.service";
 import {
   updateUserModelPreference$,
   userModelPreference,
-} from "./zero-user-data.service";
+} from "./user-data.service";
 
 const MAX_CONNECT_AGE_SECONDS = 600;
 const MAX_WEBHOOK_AGE_SECONDS = 300;
